% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResponses.R
\name{compareResponses}
\alias{compareResponses}
\title{Compare Cell Type Responses to a Reference Cell Type}
\usage{
compareResponses(
  deg_results,
  reference_cell_type,
  plot_type = "heatmap",
  similarity_metric = "jaccard"
)
}
\arguments{
\item{deg_results}{Output from compareDEGs function}

\item{reference_cell_type}{Reference cell type to compare others against}

\item{plot_type}{Type of plot ("heatmap", "barplot", "radar")}

\item{similarity_metric}{Metric to calculate response similarity}
}
\value{
A plot showing how cell types respond differentially compared to reference
}
\description{
Compare Cell Type Responses to a Reference Cell Type
}
