% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiModalIntegration.R
\name{computeModalityMixing}
\alias{computeModalityMixing}
\title{Compute modality mixing score}
\usage{
computeModalityMixing(nn_idx, modality_labels)
}
\arguments{
\item{nn_idx}{KNN indices matrix}

\item{modality_labels}{Vector of modality labels}
}
\value{
List with mixing scores
}
\description{
Compute modality mixing score
}
\keyword{internal}
