% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiModalIntegration.R
\name{computeSimpleKBET}
\alias{computeSimpleKBET}
\title{Compute simplified kBET metric}
\usage{
computeSimpleKBET(nn_idx, modality_labels)
}
\arguments{
\item{nn_idx}{KNN indices matrix}

\item{modality_labels}{Vector of modality labels}
}
\value{
List with kBET results
}
\description{
Compute simplified kBET metric
}
\keyword{internal}
