% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiModalIntegration.R
\name{computeSimpleLISI}
\alias{computeSimpleLISI}
\title{Compute simplified LISI metric}
\usage{
computeSimpleLISI(nn_idx, modality_labels)
}
\arguments{
\item{nn_idx}{KNN indices matrix}

\item{modality_labels}{Vector of modality labels}
}
\value{
List with LISI results
}
\description{
Compute simplified LISI metric
}
\keyword{internal}
