% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pseudotime.R
\name{convertSCEToCDS}
\alias{convertSCEToCDS}
\title{Convert SingleCellExperiment to Cell Data Set (CDS)}
\usage{
convertSCEToCDS(sce, reduction = "PCA", verbose = TRUE)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{reduction}{Reduction to use}

\item{verbose}{Whether to print messages}
}
\value{
A cell_data_set object
}
\description{
Convert SingleCellExperiment to Cell Data Set (CDS)
}
\keyword{internal}
