% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pseudotime.R
\name{convertSeuratToCDS}
\alias{convertSeuratToCDS}
\title{Convert Seurat Object to Cell Data Set (CDS)}
\usage{
convertSeuratToCDS(seurat_obj, reduction = "umap", verbose = TRUE)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{reduction}{Reduction to use}

\item{verbose}{Whether to print messages}
}
\value{
A cell_data_set object
}
\description{
Convert Seurat Object to Cell Data Set (CDS)
}
\keyword{internal}
