% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_richr.R
\name{createEnrichmentPlots}
\alias{createEnrichmentPlots}
\title{Create Enrichment Plots}
\usage{
createEnrichmentPlots(
  enrichment_results,
  plot_type,
  top_terms,
  separate_direction
)
}
\description{
Create Enrichment Plots
}
\keyword{internal}
