% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_richr.R
\name{createEnrichmentSummary}
\alias{createEnrichmentSummary}
\title{Create Enrichment Summary}
\usage{
createEnrichmentSummary(enrichment_results, separate_direction)
}
\description{
Create Enrichment Summary
}
\keyword{internal}
