% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{createGeneLists}
\alias{createGeneLists}
\title{Create gene lists based on grouping}
\usage{
createGeneLists(sig_degs, gene_col, by_cell_type)
}
\description{
Create gene lists based on grouping
}
\keyword{internal}
