% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_richr.R
\name{extractGenesForEnrichment}
\alias{extractGenesForEnrichment}
\title{Extract Genes for Enrichment Analysis}
\usage{
extractGenesForEnrichment(
  deg_results,
  gene_list,
  use_significant_only,
  separate_direction
)
}
\description{
Extract Genes for Enrichment Analysis
}
\keyword{internal}
