% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{filterByGroups}
\alias{filterByGroups}
\title{Filter by cell types and comparisons}
\usage{
filterByGroups(sig_degs, cell_types, comparisons)
}
\description{
Filter by cell types and comparisons
}
\keyword{internal}
