% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_richr.R
\name{performEnrichment}
\alias{performEnrichment}
\title{Perform Enrichment Analysis}
\usage{
performEnrichment(
  genes,
  annot_data,
  analysis_type,
  ontology,
  pvalue,
  qvalue,
  minGSSize,
  maxGSSize
)
}
\description{
Perform Enrichment Analysis
}
\keyword{internal}
