% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pseudotime.R
\name{plotPseudotime}
\alias{plotPseudotime}
\title{Plot Pseudotime Trajectory}
\usage{
plotPseudotime(
  object,
  color_by = "pseudotime",
  reduction = "umap",
  point_size = 0.5,
  label_groups = FALSE
)
}
\arguments{
\item{object}{Seurat or SingleCellExperiment object with pseudotime calculated}

\item{color_by}{Variable to color cells by (default: "pseudotime")}

\item{reduction}{Dimensionality reduction for plotting (default: "umap")}

\item{point_size}{Size of points (default: 0.5)}

\item{label_groups}{Whether to label groups (default: FALSE)}
}
\value{
A ggplot object
}
\description{
Create visualizations of pseudotime trajectories
}
\examples{
\dontrun{
# After calculating pseudotime
plotPseudotime(seurat_obj)
plotPseudotime(seurat_obj, color_by = "cell_type")
}

}
