% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_richr.R
\name{prepareAnnotationData}
\alias{prepareAnnotationData}
\title{Prepare Annotation Data}
\usage{
prepareAnnotationData(organism, analysis_type, build_annotation)
}
\description{
Prepare Annotation Data
}
\keyword{internal}
