% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiModalIntegration.R
\name{projectToLowDim}
\alias{projectToLowDim}
\title{Helper function to project new data to existing low dimensional space}
\usage{
projectToLowDim(new_data, ref_data, ref_embedding)
}
\arguments{
\item{new_data}{Matrix of new data in PCA space}

\item{ref_data}{Matrix of reference data in PCA space}

\item{ref_embedding}{Matrix of reference data in target low dimensional space}
}
\value{
Matrix of projected data in target low dimensional space
}
\description{
Helper function to project new data to existing low dimensional space
}
\keyword{internal}
