% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\name{scale_color_cellhet}
\alias{scale_color_cellhet}
\title{Scale Color for CellHet}
\usage{
scale_color_cellhet(...)
}
\arguments{
\item{...}{Additional arguments passed to scale_color_manual}
}
\value{
A ggplot2 scale object
}
\description{
ggplot2 color scale using CellHet default palette
}
