% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\name{scale_fill_cellhet}
\alias{scale_fill_cellhet}
\title{Scale Fill for CellHet}
\usage{
scale_fill_cellhet(...)
}
\arguments{
\item{...}{Additional arguments passed to scale_fill_manual}
}
\value{
A ggplot2 scale object
}
\description{
ggplot2 fill scale using CellHet default palette
}
