% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizePathways.R
\name{visualizePathways}
\alias{visualizePathways}
\title{Visualize Pathway Enrichment Results}
\usage{
visualizePathways(
  pathway_results,
  plot_type = "dotplot",
  top_n = 10,
  common_only = FALSE
)
}
\arguments{
\item{pathway_results}{Output from runPathwayAnalysis function}

\item{plot_type}{Type of visualization ("dotplot", "heatmap", "network")}

\item{top_n}{Number of top pathways to include per comparison}

\item{common_only}{If TRUE, only show pathways found in multiple comparisons}
}
\value{
A plot object visualizing pathway enrichment
}
\description{
Visualize Pathway Enrichment Results
}
